requirements.txt
python-telegram-bot==20.5
python-dotenv==1.0.0
# aiohttp is optional; quote endpoint uses urllib to avoid compilation issues on Windows
